Restructure and refactor to make relationships clearer.
While there are still some differences between the notes and projects
workflows, but they are now approaching alignment to the point where
parameterisation might seem to be a possibility.

Also it would seem to be much simpler to put the save and similar
db methods as methods of the mongoengine document classes.